+++ src/dreamstate/__tests__/react_tree_sync.test.ts
@@ -1,4 +1,4 @@
-import { mount } from "enzyme";
+import { mount, render } from "enzyme";
 import { createElement } from "react";
 import { act } from "react-dom/test-utils";
 
@@ -9,21 +9,23 @@ import { TestContextManager } from "@/fixtures";
 
 describe("React tree for provided and consumed components", () => {
   it("Should correctly update subscribed to functional provider elements view", () => {
-    const reactTree = mount(
+    const getRoot = () => (
       createElement(
         createProvider([ TestContextManager ]),
         {},
         createElement(function() {
           const context = useManager(TestContextManager);
 
-          return JSON.stringify(context);
-        } as any)
+          return createElement("div", {}, JSON.stringify(context));
+        })
       )
     );
 
+    expect(render(getRoot())).toMatchSnapshot();
+
+    const reactTree = mount(getRoot());
     const testContextManager: TestContextManager | null = getCurrent(TestContextManager);
 
-    expect(reactTree).toMatchSnapshot("Initial state.");
     expect(testContextManager).not.toBeNull();
 
     act(() => {
@@ -33,9 +35,8 @@ describe("React tree for provided and consumed components", () => {
     expect(testContextManager!.context.first).toBe("updated");
     expect(testContextManager!.context.second).toBe(101010);
     expect(testContextManager!.context.third).toBe(true);
+    expect(render(getRoot())).toMatchSnapshot();
 
-    reactTree.update();
-
-    expect(reactTree).toMatchSnapshot("Updated state.");
+    reactTree.unmount();
   });
 });