+++ src/dreamstate/core/provision/createProvider.ts
@@ -1,13 +1,12 @@
-import { FunctionComponent, ReactNode, useReducer } from "react";
+import { FunctionComponent, ReactNode } from "react";
 
-import { useHotObservers } from "@/dreamstate/core/observing/useHotObservers";
-import { useStaticObservers } from "@/dreamstate/core/observing/useStaticObservers";
-import { provideSubTreeRecursive } from "@/dreamstate/core/provision/provideSubTreeRecursive";
+import { createCombinedProvider } from "@/dreamstate/core/provision/createCombinedProvider";
+import { createScopedProvider } from "@/dreamstate/core/provision/createScopedProvider";
 import { ContextManager } from "@/dreamstate/core/services/ContextManager";
 import { TAnyContextManagerConstructor, TAnyObject } from "@/dreamstate/types";
 
-function forceUpdateReducer(): TAnyObject | null {
-  return {};
+export interface ICreateProviderProps {
+  isCombined?: boolean;
 }
 
 export interface IProviderProps<T> {
@@ -19,11 +18,14 @@ export interface IProviderProps<T> {
 /**
  * Create component for manual provision without HOC/Decorator-like api.
  * Useful if your root is functional component or you are using createComponent api without JSX.
+ *
+ * isCombined -> should each component work in a separate scope.
  */
 export function createProvider<
   T extends TAnyObject = TAnyObject
 >(
-  sources: Array<TAnyContextManagerConstructor>
+  sources: Array<TAnyContextManagerConstructor>,
+  { isCombined = true }: ICreateProviderProps = {}
 ): FunctionComponent<IProviderProps<T>> {
   if (!Array.isArray(sources)) {
     throw new TypeError(
@@ -37,29 +39,9 @@ export function createProvider<
     }
   }
 
-  /**
-   * Create observer component that will handle observing.
-   */
-  function Observer(props: IProviderProps<T>): ReactNode {
-    const [ , updateProviders ] = useReducer(forceUpdateReducer, null);
-
-    if (props.partialHotReplacement) {
-      useHotObservers(sources, props.initialState, updateProviders);
-    } else {
-      useStaticObservers(sources, props.initialState, updateProviders);
-    }
-
-    return provideSubTreeRecursive(props.children, sources, 0);
-  }
-
-  if (IS_DEV) {
-    Observer.displayName = `Dreamstate.Observer[${sources.map(function(it: TAnyContextManagerConstructor) {
-      return it.name;
-    })
-    }]`;
+  if (isCombined) {
+    return createCombinedProvider(sources);
   } else {
-    Observer.displayName = "DS.Observer";
+    return createScopedProvider(sources);
   }
-
-  return Observer as FunctionComponent<IProviderProps<T>>;
 }