+++ src/dreamstate/core/provision/forceUpdateReducer.ts
@@ -1,0 +1,9 @@
+import { TAnyObject } from "@/dreamstate/types";
+
+/**
+ * Reducer that returns new object every time.
+ * Used for update forcing.
+ */
+export function forceUpdateReducer(): TAnyObject | null {
+  return {};
+}