+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   * getCurrentContext moved to test-utils
   * queryData renamed to queryDataAsync
   * queryDataAsync stopped array of queries supporting
+  * mount/unmount order now matches react components
 
 ## 3.3.2 (4/11/2021)
   * 'partialHotReplacement' parameter for created provider elements that are disabled by default -> allow hot updates and partial reload of context managers